packages/blaueis-core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "blaueis-core"
version = "0.1.0"
description = "Midea HVAC serial protocol library — codec, glossary, state, query, command"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"blaueis.core" = ["data/*.yaml", "data/*.json", "data/device_quirks/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "validate_*.py"]
